refactor(lets-0np5i): extract letsconfig.MergedEnv + add launcher whitelist
Task: lets-0np5i
